+++ src/main/java/com/incidents/services/impl/RodentBaitingService.java
@@ -23,8 +23,8 @@ public class RodentBaitingService {
 	RodentBaitingDAO dao;
 	
 	@Transactional
-	public RodentBaiting create(String status, String streetAddress, String zipCode, Integer xCoordinate,
-			Integer yCoordinate, Integer ward, Integer policeDistrict, Integer communityArea, Integer latitude, Integer longitude,
+	public RodentBaiting create(String status, String streetAddress, String zipCode, Double xCoordinate,
+			Double yCoordinate, Integer ward, Integer policeDistrict, Integer communityArea, Double latitude, Double longitude,
 			String location, Integer premisesBaited, Integer premisesWithGarbage, Integer premisesWithRats, String currentActivity, String mostRecentAction) {
 		Incident newIncident = new Incident();
 		newIncident.setCreatedAt(new Date());
