+++ src/main/java/com/incidents/web/AbandonedVehiclesEndpoint.java
@@ -28,9 +28,9 @@ public class AbandonedVehiclesEndpoint {
 	  private AbandonedVehiclesService service;
 	  
 	  @RequestMapping(value = "/abandonedVehiclesEndpoint/create", method = RequestMethod.POST, headers = "Accept=application/json")
-	  public AbandonedVehicles create(Principal principal, @RequestParam("status") String status, @RequestParam("streetAddress") String streetAddress, @RequestParam("zipCode") String zipCode, @RequestParam("xCoordinate") Integer xCoordinate,
-			  @RequestParam("yCoordinate") Integer yCoordinate, @RequestParam("ward") Integer ward, @RequestParam("policeDistrict") Integer policeDistrict, @RequestParam("communityArea") Integer communityArea, @RequestParam("latitude") Integer latitude, @RequestParam("longitude") Integer longitude,
-			  @RequestParam("location") String location, @RequestParam("licensePlate") String licensePlate, @RequestParam("model") String model, @RequestParam("color") String color, @RequestParam("currentActivity") String currentActivity, @RequestParam("mostRecentAction") String mostRecentAction, @RequestParam("daysAbandoned") Integer daysAbandoned, @RequestParam("ssa") String ssa) {
+	  public AbandonedVehicles create(Principal principal, @RequestParam("status") String status, @RequestParam("streetAddress") String streetAddress, @RequestParam("zipCode") String zipCode, @RequestParam("xCoordinate") Double xCoordinate,
+			  @RequestParam("yCoordinate") Double yCoordinate, @RequestParam("ward") Integer ward, @RequestParam("policeDistrict") Integer policeDistrict, @RequestParam("communityArea") Integer communityArea, @RequestParam("latitude") Double latitude, @RequestParam("longitude") Double longitude,
+			  @RequestParam("location") String location, @RequestParam("licensePlate") String licensePlate, @RequestParam("model") String model, @RequestParam("color") String color, @RequestParam("currentActivity") String currentActivity, @RequestParam("mostRecentAction") String mostRecentAction, @RequestParam("daysAbandoned") Double daysAbandoned, @RequestParam("ssa") String ssa) {
 		  MyUser user;
 		  
 		    if (principal != null) {
